Create CordovaExtension
The CordovaExtension is designed to work as the loader and the
manager of cordova plugins, the goal is to keep the implementation
of cordova plugin's native and js source codes remain the same
without any change, and works fine in this extension.

What we've done so far:
1. Imported part of the cordova data structures, both native and js
part, in order to keep the plugin interface remains the same;
2. Created CordovaPluginDeviceDemo project to host the demo app and
Device plugin implementation as the test case;
3. Small changes has made on CrosswalkLite library to make the extension
build and leverage the function of Invocation.

For now the cordova message can be transferred to native side, and the
Device plugin can be invoked and handle the message.