Removed unnecessary private properties in XWalkChannel
For the properties with public read and private write access, just
declared them as:
  private(set) public var foo: ...
no need to use a private var property with a public calculate property.

Also fixed a dispatch_async calling issue, we should not use 'self'
inside an async block, as it may hold the self reference and generate
a retain cycle. Instead we should pass a weak reference into the block
and try to hold a strong reference and then use it.