Added XWalkViewTest target for Unit Testing.
Use build-in XCTest as our unit test framework, and generated the
stub test codes for classes in XWalkView. We are intended to cover
every public function in each class.

Use 'Product -> Test' to run the unit test, or install xctool with
'brew install xctool'
and run the xctool as:
'xctool -scheme XWalkViewTests -sdk iphonesimulator test'
to see the result in commandline.