Use derived XWalkView instead of WKWebView extension.
Because the associated object's deinit time is not as expected,
we decide to choose derived XWalkView instead of WKWebView extension.
XWalkView will hold the channels and manage the life cycle of them;
Channel will weak reference the thread object, and the loadExtension
interface won't need to ask user about the threading.

Testcases are upgraded accordingly.