Read extension configuration from extensions.plist
This feature will help the extensions which are published through
CocoaPods to define the JavaScript namespace to native class
mapping, as we cannot modify the Info.plist by the definitions
of podfile.

If the extension provides the extensions.plist, we'll read the
mapping from it; otherwise we will read them from the Info.plist.